+++ packages/components/CHANGELOG.md
@@ -12,11 +12,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 #### Added
 
 - `underline` prop for `Link Pure`
-- `Checkbox Wrapper`, `Tag Dismissible` supports `theme` prop
+- `Checkbox Wrapper`, `Radio Button Wrapper` and `Tag Dismissible` supports `theme` prop
 
 #### Changed
 
-- `Spinner`, `Link Pure`, `Button Pure`, `Checkbox Wrapper`, `Tag Dismissible` matches new design language
+- `Spinner`, `Link Pure`, `Button Pure`, `Checkbox Wrapper`, `Radio Button Wrapper`, `Tag Dismissible` matches new design language
 - Prop `weight` for `Link Pure` and `Button Pure` is deprecated now, only regular font weight will be applied
 
 ```diff